fix: Address all analyzer rule violations
- Remove unused variables and fields across codebase (IDE0059, IDE0052)
- Simplify collection initializations using collection expressions (IDE0028)
- Mark QueryEnumerator members as readonly (IDE0251)
- Update to System.Threading.Lock for .NET 10 (IDE0330)
- Remove redundant casts (IDE0004)
- Optimize event benchmark handlers

Fixes 27 analyzer violations across source, benchmarks, and tests:
- Source code: 10 fixes (generators, graphics, spatial, core)
- Benchmarks: 2 fixes (event and command buffer benchmarks)
- Tests: 15 fixes (generator, spatial, core tests)

All builds now pass with 0 warnings and 0 errors.

+++ tests/KeenEyes.Core.Tests/RemoveComponentTests.cs
@@ -360,7 +360,7 @@ public class RemoveComponentTests
         {
             world.Add(entity, new TestPosition { X = i, Y = i });
             Assert.True(world.Has<TestPosition>(entity));
-            Assert.Equal((float)i, world.Get<TestPosition>(entity).X);
+            Assert.Equal(i, world.Get<TestPosition>(entity).X);
 
             Assert.True(world.Remove<TestPosition>(entity));
             Assert.False(world.Has<TestPosition>(entity));
